Fix multiple role selection error
When multiple roles are selected, it triggers TWO on_user_event errors, where the second one is empty. this throws an error.
To fix this, I changed the handling to only allow one selection at a time, which maintains the functionality, but disallows for multiple events to trigger the listener.