Update reaction_roles.toml
Added a section to reaction roles for users to choose their own color based on all the new color role IDs.  As far as I can tell the dropdown menu is just read from this toml file when the embed is created.  If something needs to be added to the actual cog, I haven't done that...