Clean Up of general_trans and pronunciation feature
Clean up includes fixing a bit of the imports to look cleaner, adding the whitelist as a .txt file to the bot to read instead of hardcoding it and taking up so many lines of code. This can be moved to the DB if you prefer..., LANGUAGES is a dictionary we import from the google_trans module so there is no real need to add it to the class' namespace, changed the capitalization of the output of Language Names, and adjusted the confidence percentage to have only two decimal places, and output the list of language codes in 3 columns of text instead of 1 long one. (This might look bad on mobile... I am not sure).

I have commented three places where I am unsure if the code is necessary/good practice: 1) the self.threshold variable is never used... 2) The bot does not respond to messages shorter than 7 characters... we do not let it translate short words?  and 3) I am not confident the detection of "en" is done the best way.

Finally I have added pronunciation to the footer for languages that provide it.