Update general_trans.py
Made a static method of getting the pronunciation so the code wasn't repeated for both the listening command and the slash command.

Also improved the pronunciation getting method after noticing some odd behaviour during testing.

Added Crambor's suggestion of not outputting a translation if the original is the same as the translated text, to avoid unwanted spam translations.

Finally made uniform the variable name for the translated data:  It was "translated" in one command and "translation" in another. I just made both "translation"