Update general_latex.py
Added two helper slash-commands to show the user available colors for BG and Text.  These just pull up a stored image from utilities/cog_helpers and shows it to the user in an ephemeral message.

Also transferred the colors list to a .json file and reads those values into the code.

Finally, made the failure message for bad latex ephemeral to lessen the embarrassment of making a mistake with the /latex command.

FUTURE GOALS:
I would like to have the args for /latex come with a description.  This should be possible... but I cannot figure it out.