Update utility_run_code.py
Added functionality for input() by requesting the user add those values after the codeblock in order. As a result of this, the regex (re) module was added.
for example:
/run
\`\`\`py
x = input()
y = input()
print(f"{x} and {y}")
\`\`\`
apples
bananas

Also moved the embed creation to a separate method to avoid repetitive code. Each time you need an embed you build it with the class method, sending the arguments for name, value, and is_error.  Indicating True for is_error will result in a red embed with "Oops..." as the title whereas leaving it out of the arguments or sending False will result in a green embed with "Python 3.10" for the title.

Finally, attempted to correct the odd iOS quotes using escape characters inside of the quotes instead of the doc string method since that was not working.