Update utility_run_code.py
Fixes after debugging with Zorak in the dev server.

It seemed str.translate(str.maketrans()) was not going to work with strange iOS quotes so I replaced it with a loop of str.replace()s and now it works.  Could potentially behave strangely with quotes from other operating systems, but iPhone is fine.

Changed some formatting for some of the error embeds to look nicer.