Added editable /run commands
Changed logging_message_edit.py to look for changes to a /run command and re-run it as necessary.  This involved storing the previous message id as an attribute in the utility_run_code.py UtilityRunCode class so the old embed can be deleted.

***WARNING***
I am unsure if the code I've added to logging_message_edit.py negatively affects the logging of message changes that had previously worked.  It should be ok... but should be tested before pulling.

Also as an added bonus I made Zorak's dialogue a bit sassier (Zadok's suggestion) to attempt to discourage using input() but still allowing it.

Finally. fixed a small error in the replace() function for handling input().  Only one instance of the input() should be replaced per argument, and as it was before, if the user had identical prompts for all of their inputs, it would change all of them to one argument's value instead of just one.