Create general_mybot.py
A fun feature that uses an API to give a random picture of a robot/cat/human/monster based on a provided text seed (optional).

If text seed is not provided, the user's name is used for the API.

There is also an optional prompt to select a specific type (robot,monster,robot head,cat,human), if left blank it will always be robot.