Update general_coinflip.py
Attempted to redirect all gambling to the Off Topic channel with a ping for the user.

Could not test fully because I cannot connect to the db, but The message saying I do not have any points (because my wallet wasn't found in the db) was directed to Off Topic.  So I think the rest of the messages in this cog will work similarly.